move other module files to ufm dir
